third_party/blink/web_tests/wpt_internal/bluetooth/characteristic/notifications/event-after-starting.https.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<script src="/bluetooth/resources/bluetooth-test.js"></script>
<script src="/bluetooth/resources/bluetooth-fake-devices.js"></script>
<script>
'use strict';
bluetooth_test(() => {
  return setBluetoothFakeAdapter('HeartRateAdapter')
    .then(() => requestDeviceWithTrustedClick({
      filters: [{services: ['heart_rate']}]}))
    .then(device => device.gatt.connect())
    .then(gattServer => gattServer.getPrimaryService('heart_rate'))
    .then(service => service.getCharacteristic('heart_rate_measurement'))
    .then(characteristic => {
      return assert_promise_resolves_before_event(
        characteristic, 'startNotifications', 'characteristicvaluechanged')
        .then(() => characteristic.stopNotifications())
        .then(() => assert_no_events(characteristic,
                                     'characteristicvaluechanged'));
    });
}, 'Starting notifications should fire an event and we should get no events ' +
   'after notifications are stopped.');
</script>
